backend/pyproject.toml
[project]
name = "coxwave-openai-hackerton"
version = "0.1.0"
description = "Video generation AI agent system for marketing videos"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.6.9",
    "fastapi>=0.114.1,<0.116",
    "uvicorn[standard]>=0.36,<0.37",
    "openai>=1.40",
    "openai-chatkit>=1.4.0,<2",
    "jinja2>=3.0",
]

[tool.uv]
dev-dependencies = [
    "pyright>=1.1.400",
    "pre-commit>=4.0.0",
]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "basic"
